[Feat/#82] naver, google login test

+++ src/main/java/com/example/waggle/global/security/oauth2/info/KakaoOAuth2User.java
@@ -3,11 +3,11 @@ package com.example.waggle.global.security.oauth2.info;
 import java.util.Map;
 
 public class KakaoOAuth2User extends OAuth2UserInfo{
-    private Integer id;
+    private Long id;
 
     public KakaoOAuth2User(Map<String, Object> attributes) {
         super((Map<String, Object>) attributes.get("kakao_account"));
-        this.id = (Integer) attributes.get("id");
+        this.id = (Long) attributes.get("id");
     }
 
     @Override
